+++ docs/CHANGE_LOG.md
@@ -1,6 +1,7 @@
 # Reattribute.js - Change Log:
 
 ## Version 1.2.0:
+- Added support to set attribute values directly from function call results!
 - Improved the ignore check to use actual boolean value.
 - Added improved checks around the elements (incase they are removed from the DOM).
 - Added a new configuration option called "assignMissingIds" (defaults to false), which will assign Ids to the elments that do not contain them.